handleGetCurrentUserをカスタムフックに切り出し
