+++ frontend/src/components/pages/Welcome.tsx
@@ -1,51 +1,13 @@
-import { VFC, memo, useState } from 'react'
-import { useHistory, useLocation } from 'react-router-dom'
+import { VFC, memo } from 'react'
 
 import { Box, Button, Flex, Image, Spinner, Text } from '@chakra-ui/react'
-import { signInWithPopup, GoogleAuthProvider } from 'firebase/auth'
 
 import googleIcon from 'assets/images/google_icon.svg'
 import LogoWithCopy from 'assets/images/logo-with-copy.svg'
-import { getCurrentUser } from 'lib/api/session'
-import { auth } from 'lib/firebase'
+import { useSignInWithGoogle } from 'hooks/useSignInWithGoogle'
 
 export const Welcome: VFC = memo(() => {
-  const [isLoading, setIsLoading] = useState(false)
-  const history = useHistory()
-  const { search } = useLocation()
-  const query = new URLSearchParams(search)
-  const invitationToken = query.get('invitation_token')
-
-  const handleGetCurrentUser = async () => {
-    const token = await auth.currentUser?.getIdToken(true)
-    const res = await getCurrentUser(token)
-    return res
-  }
-
-  const signInWithGoogle = () => {
-    const provider = new GoogleAuthProvider()
-    signInWithPopup(auth, provider)
-      .then(() => {
-        setIsLoading(true)
-      })
-      .then(handleGetCurrentUser)
-      .then((res) => {
-        if (res?.data.isExisted) {
-          history.push('/')
-        } else {
-          history.push({
-            pathname: '/onboarding',
-            state: { invitationToken },
-          })
-        }
-      })
-      .catch((error) => {
-        console.log(error)
-      })
-      .finally(() => {
-        setIsLoading(false)
-      })
-  }
+  const { signInWithGoogle, isLoading } = useSignInWithGoogle()
 
   return (
     <>
